:recycle: move tests in same file as per new standard
