+++ marketplace-domain/src/contribution/projectors/mod.rs
@@ -9,6 +9,3 @@ pub use project::ProjectProjector;
 
 mod contributor;
 pub use contributor::ContributorProjector;
-
-#[cfg(test)]
-mod contributor_test;
